Only generate random_key_file if needed
Previously the random_key_file.h was generated every time. There was no
check that checked if the file already existed. This check has now been
introduced. When building we first check the folder
RANDOM_KEY_SOURCE_DIRECTORY supplied via input to the build system.
If the file was not located there it will look in the build folder.
If random_key_file.h was not located there it will be generated.