LIC-704 Separating the input of packagetool from the output of packagetool when creating test_library
Currently, the input and output are read from/written to the same location (but separated in time). Moving the packagetool input into build dir subdirectory packagetool_input, and leaving the packagetool output in build dir. This simplifies maintenance as running the packagetool command on the input does not accidentally run it on the output, depending on when the command is run.